requirements.txt
PyYAML==3.11
flask==0.10.1
Flask-Script==2.0.5
